Fixed history update bug.
The problem was the call to update history was being triggered before the call to set the contract had settled.  Because of this, when the call to set the history came through, the function would return immediately because the contract had not yet arrived.

By debouncing the call to updateHIstory, we allow the call to set the contract time to settle before triggering the history update